Use go embed to extract hardcoded strings to external files
