+++ test/new-e2e/language-detection/node_test.go
@@ -6,28 +6,15 @@
 package languagedetection
 
 import (
+	_ "embed"
 	"fmt"
 	"strings"
 
 	"github.com/stretchr/testify/require"
 )
 
-const nodeProg = `
-const http = require('http');
-
-const hostname = '127.0.0.1';
-const port = 3000;
-
-const server = http.createServer((req, res) => {
-  res.statusCode = 200;
-  res.setHeader('Content-Type', 'text/plain');
-  res.end('Hello World');
-});
-
-server.listen(port, hostname, () => {
-  console.log('Server running');
-});
-`
+//go:embed etc/node_server.js
+var nodeProg string
 
 func (s *languageDetectionSuite) installNode() {
 	s.Env().VM.Execute(
